Added file via upload
This was my first script for this class and deep learning in general. It's very interesting and fun! I plan on doing much more soon, the logistics problem was especially cool to me because I feel like it is easier to think of direct applications for it. This is a .py file instead of the initial .ipynb file in colab for easier use. This is also why its structure is a bit different compared to some of my other code.